+++ frontend/src/main.ts
@@ -3,7 +3,7 @@ import './app.css'
 import App from './App.svelte'
 
 console.log(
-  '%c🚜 TraktOr%c\nDer Traktor rollt an. Guten Tag, Entwickler :)',
+  '%c🚜 [trakt]Or%c\nDer Traktor rollt an. Guten Tag, neugieriger Entdecker :)',
   'font-weight: 700; font-size: 14px;',
   'font-weight: 400;'
 )